feat: split drawer memo modes
