+++ electron/preload.js
@@ -90,6 +90,8 @@ const electronAPI = {
     ipcRenderer.invoke("ws:create-project", { workspacePath, name, id }),
   wsSelectDirectory: () => ipcRenderer.invoke("ws:select-directory"),
   wsGetLegacyProjects: () => ipcRenderer.invoke("ws:get-legacy-projects"),
+  wsExportLegacyProjects: (workspacePath) =>
+    ipcRenderer.invoke("ws:export-legacy-projects", { workspacePath }),
   wsMigrateProjects: (workspacePath) =>
     ipcRenderer.invoke("ws:migrate-projects", { workspacePath }),
 };
