+++ package.json
@@ -79,6 +79,10 @@
     "@commonify/lowdb": "^3.0.0",
     "electron-log": "^5.4.3",
     "lodash": "^4.18.1",
-    "marked": "^18.0.2"
+    "marked": "^18.0.2",
+    "quill": "^2.0.3",
+    "quill-delta-to-html": "^0.12.1",
+    "turndown": "^7.2.1",
+    "turndown-plugin-gfm": "^1.0.2"
   }
 }
