+++ src/components/MigrationWizard.svelte
@@ -7,12 +7,11 @@
   export let toggle: () => void;
 
   type LegacyProject = { id: string; name: string; taskCount: number };
-  type MigrateResult = {
+  type ExportResult = {
     success: boolean;
     migrated: { name: string; count: number }[];
     errors: { name: string; error: string }[];
   };
-
   type Phase = "idle" | "loading" | "ready" | "running" | "done";
 
   let phase: Phase = "idle";
@@ -20,7 +19,7 @@
   let selectedPath = $workspace_store.activeWorkspacePath ?? "";
   let customPath = "";
   let useCustom = false;
-  let result: MigrateResult | null = null;
+  let result: ExportResult | null = null;
   let loadError = "";
 
   $: targetPath = useCustom ? customPath : selectedPath;
@@ -32,7 +31,7 @@
     try {
       legacyProjects = await platform.wsGetLegacyProjects();
       phase = legacyProjects.length === 0 ? "idle" : "ready";
-      if (legacyProjects.length === 0) loadError = "移行対象のプロジェクトがありません。";
+      if (legacyProjects.length === 0) loadError = "No db.json projects found.";
     } catch (e) {
       loadError = String(e);
       phase = "idle";
@@ -47,11 +46,11 @@
     }
   }
 
-  async function handleMigrate() {
+  async function handleExport() {
     if (!targetPath) return;
     phase = "running";
     try {
-      result = await platform.wsMigrateProjects(targetPath);
+      result = await platform.wsExportLegacyProjects(targetPath);
       if (result) {
         await workspace_store.refreshProjects();
       }
@@ -73,83 +72,82 @@
 
 <Modal {show} toggle={handleClose} width="44rem" height="auto">
   <div class="container">
-    <div class="header">レガシーデータを移行</div>
+    <div class="header">Export db.json projects</div>
 
     <div class="body">
       {#if phase === "loading"}
-        <p class="note">読み込み中...</p>
+        <p class="note">Loading...</p>
       {:else if phase === "idle"}
         {#if loadError}
           <p class="empty-note">{loadError}</p>
         {/if}
       {:else if phase === "ready" || phase === "running"}
-        <!-- Project list -->
-        <p class="section-label">移行されるプロジェクト ({legacyProjects.length} 件)</p>
+        <p class="section-label">Projects to export ({legacyProjects.length})</p>
         <ul class="project-list">
           {#each legacyProjects as p (p.id)}
             <li class="project-item">
               <span class="project-name">{p.name}</span>
-              <span class="task-count">{p.taskCount} タスク</span>
+              <span class="task-count">{p.taskCount} tasks</span>
             </li>
           {/each}
         </ul>
 
-        <!-- Target workspace -->
-        <p class="section-label">移行先ワークスペース</p>
+        <p class="section-label">Destination workspace</p>
         <div class="target-area">
           {#if workspaces.length > 0}
             <select
               class="ws-select"
               bind:value={selectedPath}
               on:change={() => (useCustom = false)}
+              disabled={phase === "running"}
             >
               {#each workspaces as ws (ws.path)}
                 <option value={ws.path}>{ws.label}</option>
               {/each}
             </select>
-            <span class="or-text">または</span>
+            <span class="or-text">or</span>
           {/if}
-          <button class="select-btn" on:click={handleSelectCustom}> フォルダを選択... </button>
+          <button class="select-btn" disabled={phase === "running"} on:click={handleSelectCustom}>
+            Select folder...
+          </button>
           {#if useCustom && customPath}
             <span class="custom-path">{customPath}</span>
           {/if}
         </div>
 
-        <p class="warn-note">
-          移行前に <code>db.json.bak</code> としてバックアップが作成されます。
-        </p>
+        <p class="warn-note">This exports Markdown files only. The source db.json is not changed.</p>
+        {#if phase === "running"}
+          <p class="note">Exporting...</p>
+        {/if}
       {:else if phase === "done" && result}
-        <!-- Results -->
         {#if result.migrated.length > 0}
-          <p class="section-label">移行成功 ({result.migrated.length} 件)</p>
+          <p class="section-label">Exported ({result.migrated.length})</p>
           <ul class="result-list">
             {#each result.migrated as m}
-              <li class="result-ok">✓ {m.name} ({m.count} タスク)</li>
+              <li class="result-ok">OK: {m.name} ({m.count} tasks)</li>
             {/each}
           </ul>
         {/if}
         {#if result.errors.length > 0}
-          <p class="section-label error-label">エラー ({result.errors.length} 件)</p>
+          <p class="section-label error-label">Errors ({result.errors.length})</p>
           <ul class="result-list">
             {#each result.errors as e}
-              <li class="result-err">✗ {e.name}: {e.error}</li>
+              <li class="result-err">Error: {e.name}: {e.error}</li>
             {/each}
           </ul>
         {/if}
         {#if result.errors.length === 0}
-          <p class="success-note">移行が完了しました。</p>
+          <p class="success-note">Export completed.</p>
         {/if}
       {/if}
     </div>
 
     <div class="footer">
       {#if phase === "ready"}
-        <button class="migrate-btn" disabled={!targetPath} on:click={handleMigrate}>
-          移行する
-        </button>
+        <button class="export-btn" disabled={!targetPath} on:click={handleExport}>Export</button>
       {/if}
       <button class="close-btn" on:click={handleClose}>
-        {phase === "done" ? "閉じる" : "キャンセル"}
+        {phase === "done" ? "Close" : "Cancel"}
       </button>
     </div>
   </div>
@@ -164,6 +162,7 @@
     border-radius: 0.5rem;
     overflow: hidden;
   }
+
   .header {
     padding: 0.75rem 1rem;
     font-weight: bold;
@@ -172,6 +171,7 @@
     background-color: var(--theme-color-Main-main);
     border-bottom: 1px solid var(--theme-color-Sub-dark);
   }
+
   .body {
     padding: 1rem;
     display: flex;
@@ -180,115 +180,109 @@
     max-height: 60vh;
     overflow-y: auto;
   }
+
   .section-label {
     font-size: 0.9rem;
     font-weight: bold;
     color: var(--theme-color-Sub-main);
     margin: 0.5rem 0 0.25rem;
-    opacity: 0.7;
+    opacity: 0.75;
   }
-  .error-label {
-    color: var(--theme-color-Error-main);
-    opacity: 1;
-  }
+
   .project-list,
   .result-list {
     list-style: none;
-    margin: 0;
     padding: 0;
+    margin: 0;
     display: flex;
     flex-direction: column;
-    gap: 0.2rem;
+    gap: 0.25rem;
   }
+
   .project-item {
     display: flex;
+    justify-content: space-between;
     align-items: center;
-    padding: 0.35rem 0.75rem;
+    padding: 0.45rem 0.65rem;
     border-radius: 0.25rem;
     background-color: var(--theme-color-Main-main);
-    gap: 0.5rem;
+    gap: 0.75rem;
   }
+
   .project-name {
-    flex: 1;
     color: var(--theme-color-Sub-main);
-    font-size: 0.95rem;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
   }
-  .task-count {
-    font-size: 0.8rem;
+
+  .task-count,
+  .note,
+  .warn-note,
+  .empty-note {
     color: var(--theme-color-Sub-dark);
-    flex-shrink: 0;
+    font-size: 0.85rem;
   }
-  .result-ok {
-    font-size: 0.9rem;
-    color: var(--theme-color-Success-main);
-    padding: 0.2rem 0.5rem;
-  }
-  .result-err {
-    font-size: 0.9rem;
-    color: var(--theme-color-Error-main);
-    padding: 0.2rem 0.5rem;
-    word-break: break-all;
-  }
+
   .target-area {
     display: flex;
     flex-wrap: wrap;
     align-items: center;
     gap: 0.5rem;
   }
+
+  .ws-select,
+  .select-btn,
+  .export-btn,
+  .close-btn {
+    border: none;
+    border-radius: 0.25rem;
+    padding: 0.35rem 0.75rem;
+    font-size: 0.9rem;
+  }
+
   .ws-select {
     background-color: var(--theme-color-Main-dark);
     color: var(--theme-color-Sub-main);
     border: 1px solid var(--theme-color-Sub-dark);
-    border-radius: 0.25rem;
-    padding: 0.3rem 0.5rem;
-    font-size: 0.9rem;
-    min-width: 10rem;
   }
-  .or-text {
-    font-size: 0.85rem;
-    color: var(--theme-color-Sub-dark);
-  }
-  .select-btn {
+
+  .select-btn,
+  .export-btn {
     cursor: pointer;
-    border: none;
-    border-radius: 0.25rem;
-    padding: 0.3rem 0.75rem;
-    font-size: 0.9rem;
-    background-color: var(--theme-color-Theme-main);
+    background-color: var(--theme-color-Primary-main);
     color: white;
   }
-  .select-btn:hover {
-    background-color: var(--theme-color-Theme-dark);
+
+  .select-btn:disabled,
+  .export-btn:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
   }
+
   .custom-path {
-    font-size: 0.8rem;
-    color: var(--theme-color-Sub-main);
     flex: 1;
+    min-width: 12rem;
     overflow: hidden;
     text-overflow: ellipsis;
     white-space: nowrap;
+    color: var(--theme-color-Sub-main);
+    font-size: 0.85rem;
   }
-  .warn-note {
-    font-size: 0.82rem;
-    color: var(--theme-color-Sub-dark);
-    margin-top: 0.25rem;
-  }
-  .warn-note code {
-    background-color: var(--theme-color-Main-dark);
-    padding: 0.1rem 0.3rem;
-    border-radius: 0.2rem;
-    font-size: 0.82rem;
-  }
-  .success-note {
-    font-size: 0.9rem;
+
+  .result-ok {
     color: var(--theme-color-Success-main);
-    margin-top: 0.25rem;
   }
-  .note,
-  .empty-note {
-    font-size: 0.9rem;
-    color: var(--theme-color-Sub-dark);
+
+  .result-err,
+  .error-label {
+    color: var(--theme-color-Error-main);
   }
+
+  .success-note {
+    color: var(--theme-color-Success-main);
+  }
+
   .footer {
     display: flex;
     justify-content: flex-end;
@@ -297,32 +291,10 @@
     border-top: 1px solid var(--theme-color-Sub-dark);
     background-color: var(--theme-color-Main-main);
   }
-  .migrate-btn {
-    cursor: pointer;
-    border: none;
-    border-radius: 0.25rem;
-    padding: 0.3rem 1rem;
-    font-size: 0.9rem;
-    background-color: var(--theme-color-Primary-main);
-    color: white;
-  }
-  .migrate-btn:hover:not(:disabled) {
-    background-color: var(--theme-color-Primary-dark);
-  }
-  .migrate-btn:disabled {
-    opacity: 0.4;
-    cursor: not-allowed;
-  }
+
   .close-btn {
     cursor: pointer;
-    border: none;
-    border-radius: 0.25rem;
-    padding: 0.3rem 1rem;
-    font-size: 0.9rem;
     background-color: var(--theme-color-Sub-dark);
     color: var(--theme-color-Main-main);
-  }
-  .close-btn:hover {
-    opacity: 0.8;
   }
 </style>
