+++ src/types/app.ts
@@ -94,6 +94,11 @@ export interface ElectronAPI {
   ) => Promise<{ success: boolean; projectDir?: string; dirName?: string; error?: string }>;
   wsSelectDirectory: () => Promise<string | null>;
   wsGetLegacyProjects: () => Promise<{ id: string; name: string; taskCount: number }[]>;
+  wsExportLegacyProjects: (workspacePath: string) => Promise<{
+    success: boolean;
+    migrated: { name: string; count: number }[];
+    errors: { name: string; error: string }[];
+  }>;
   wsMigrateProjects: (workspacePath: string) => Promise<{
     success: boolean;
     migrated: { name: string; count: number }[];
