+++ tests/component/Memo.test.js
@@ -2,12 +2,102 @@ import { fireEvent, render, waitFor } from "@testing-library/svelte";
 import { tick } from "svelte";
 import { vi } from "vitest";
 
+const quillInstances = [];
+
+vi.mock("quill", () => {
+  class MockQuill {
+    constructor(element, options) {
+      this.element = element;
+      this.options = options;
+      this.root = document.createElement("div");
+      this.root.className = "ql-editor";
+      this.theme = { tooltip: { hide: vi.fn() } };
+      this.handlers = {};
+      this.on = vi.fn((event, callback) => {
+        this.handlers[event] = callback;
+      });
+      this.enable = vi.fn();
+      this.setContents = vi.fn();
+      this.setText = vi.fn();
+      this.getContents = vi.fn(() => ({ ops: [{ insert: "changed\n" }] }));
+      this.getLength = vi.fn(() => 0);
+      this.getSelection = vi.fn(() => null);
+      this.hasFocus = vi.fn(() => false);
+      this.setSelection = vi.fn();
+      this.deleteText = vi.fn();
+      this.insertText = vi.fn();
+      this.element?.appendChild(this.root);
+      quillInstances.push(this);
+    }
+  }
+  return { default: MockQuill };
+});
+
 import Memo from "../../src/components/Memo.svelte";
 
-describe("Memo - view mode (default)", () => {
+function renderMarkdownMemo(props = {}) {
+  return render(Memo, {
+    props: {
+      saveMemo: vi.fn(),
+      content: "",
+      isWorkspaceProject: true,
+      ...props,
+    },
+  });
+}
+
+describe("Memo mode routing", () => {
+  beforeEach(() => {
+    quillInstances.length = 0;
+    window.electronAPI = { wsResolveMemoAsset: vi.fn(), openExternalLink: vi.fn() };
+  });
+
+  afterEach(() => {
+    delete window.electronAPI;
+  });
+
+  test("uses Quill for db.json Projects even when a workspace path exists", () => {
+    const saveMemo = vi.fn();
+    render(Memo, {
+      props: {
+        saveMemo,
+        content: { ops: [{ insert: "legacy\n" }] },
+        workspaceProjectDir: "C:\\workspace\\project",
+        taskId: "task-1",
+        isWorkspaceProject: false,
+      },
+    });
+
+    expect(quillInstances).toHaveLength(1);
+    expect(document.querySelector(".cm-editor")).not.toBeInTheDocument();
+    expect(quillInstances[0].setContents).toHaveBeenCalledWith({ ops: [{ insert: "legacy\n" }] });
+  });
+
+  test("uses Markdown editor for workspace projects", () => {
+    renderMarkdownMemo({ content: "# Workspace" });
+
+    expect(quillInstances).toHaveLength(0);
+    expect(document.querySelector(".preview-mode")).toBeInTheDocument();
+    expect(document.querySelector(".preview h1")).toHaveTextContent("Workspace");
+  });
+
+  test("db.json Projects save Quill Delta content", async () => {
+    const saveMemo = vi.fn();
+    render(Memo, { props: { saveMemo, content: "", isWorkspaceProject: false } });
+    const quill = quillInstances.at(-1);
+
+    quill.handlers["text-change"]({}, {}, "user");
+    await tick();
+
+    expect(saveMemo).toHaveBeenCalledWith({ ops: [{ insert: "changed\n" }] });
+  });
+});
+
+describe("Markdown Memo - view mode", () => {
   let saveMemo;
 
   beforeEach(() => {
+    quillInstances.length = 0;
     saveMemo = vi.fn();
     window.electronAPI = { wsResolveMemoAsset: vi.fn(), openExternalLink: vi.fn() };
   });
@@ -17,30 +107,28 @@ describe("Memo - view mode (default)", () => {
   });
 
   test("renders in view mode by default (no CM6 editor)", () => {
-    render(Memo, { props: { saveMemo, content: "" } });
+    renderMarkdownMemo({ saveMemo, content: "" });
     expect(document.querySelector(".cm-editor")).not.toBeInTheDocument();
     expect(document.querySelector(".preview-mode")).toBeInTheDocument();
   });
 
   test("shows placeholder when content is empty", () => {
-    render(Memo, { props: { saveMemo, content: "" } });
+    renderMarkdownMemo({ saveMemo, content: "" });
     expect(document.querySelector(".placeholder")).toBeInTheDocument();
   });
 
   test("renders markdown content as HTML in view mode", () => {
-    render(Memo, { props: { saveMemo, content: "# Hello\n\nWorld" } });
+    renderMarkdownMemo({ saveMemo, content: "# Hello\n\nWorld" });
     const preview = document.querySelector(".preview");
     expect(preview).toBeInTheDocument();
     expect(preview.querySelector("h1")).toHaveTextContent("Hello");
   });
 
   test("renders wiki links with resolved state when memo exists", () => {
-    render(Memo, {
-      props: {
-        saveMemo,
-        content: "[[Daily Notes]] and [[Missing Note]]",
-        memoTitles: ["Daily Notes"],
-      },
+    renderMarkdownMemo({
+      saveMemo,
+      content: "[[Daily Notes]] and [[Missing Note]]",
+      memoTitles: ["Daily Notes"],
     });
 
     const links = document.querySelectorAll(".preview .wiki-link");
@@ -55,13 +143,11 @@ describe("Memo - view mode (default)", () => {
       url: "file:///C:/workspace/project/task-1/assets/diagram.png",
     });
 
-    render(Memo, {
-      props: {
-        saveMemo,
-        content: "![Diagram](./assets/diagram.png)",
-        workspaceProjectDir: "C:\\workspace\\project",
-        taskId: "task-1",
-      },
+    renderMarkdownMemo({
+      saveMemo,
+      content: "![Diagram](./assets/diagram.png)",
+      workspaceProjectDir: "C:\\workspace\\project",
+      taskId: "task-1",
     });
 
     await waitFor(() => {
@@ -73,24 +159,25 @@ describe("Memo - view mode (default)", () => {
     });
   });
 
-  test("converts legacy Quill Delta object to readable markdown text", () => {
+  test("converts legacy Quill Delta object to readable markdown text in workspace view", () => {
     const delta = { ops: [{ insert: "hello" }, { insert: "\nworld" }] };
-    render(Memo, { props: { saveMemo, content: delta } });
+    renderMarkdownMemo({ saveMemo, content: delta });
     expect(document.querySelector(".preview")).toHaveTextContent("hello");
     expect(document.querySelector(".preview")).toHaveTextContent("world");
     expect(document.querySelector(".preview")).not.toHaveTextContent('"ops"');
   });
 
   test("readOnly: no placeholder shown when empty", () => {
-    render(Memo, { props: { saveMemo, content: "", readOnly: true } });
+    renderMarkdownMemo({ saveMemo, content: "", readOnly: true });
     expect(document.querySelector(".placeholder")).not.toBeInTheDocument();
   });
 });
 
-describe("Memo - edit mode", () => {
+describe("Markdown Memo - edit mode", () => {
   let saveMemo;
 
   beforeEach(() => {
+    quillInstances.length = 0;
     saveMemo = vi.fn();
     window.electronAPI = { wsSaveMemoImage: vi.fn() };
   });
@@ -100,7 +187,7 @@ describe("Memo - edit mode", () => {
   });
 
   test("clicking edit mode switch shows CM6 editor", async () => {
-    render(Memo, { props: { saveMemo, content: "hello" } });
+    renderMarkdownMemo({ saveMemo, content: "hello" });
     await fireEvent.click(document.querySelector(".edit-mode-btn"));
     await waitFor(() => {
       expect(document.querySelector(".cm-editor")).toBeInTheDocument();
@@ -108,7 +195,7 @@ describe("Memo - edit mode", () => {
   });
 
   test("edit mode shows read mode switch", async () => {
-    render(Memo, { props: { saveMemo, content: "hello" } });
+    renderMarkdownMemo({ saveMemo, content: "hello" });
     await fireEvent.click(document.querySelector(".edit-mode-btn"));
     await waitFor(() => {
       expect(document.querySelector(".read-mode-btn")).toBeInTheDocument();
@@ -116,7 +203,7 @@ describe("Memo - edit mode", () => {
   });
 
   test("clicking read mode switch returns to view mode", async () => {
-    render(Memo, { props: { saveMemo, content: "hello" } });
+    renderMarkdownMemo({ saveMemo, content: "hello" });
     await fireEvent.click(document.querySelector(".edit-mode-btn"));
     await waitFor(() => expect(document.querySelector(".read-mode-btn")).toBeInTheDocument());
 
@@ -128,7 +215,7 @@ describe("Memo - edit mode", () => {
   });
 
   test("readOnly: clicking does not enter edit mode", async () => {
-    render(Memo, { props: { saveMemo, content: "hello", readOnly: true } });
+    renderMarkdownMemo({ saveMemo, content: "hello", readOnly: true });
     await fireEvent.click(document.querySelector(".preview-mode"));
     await tick();
     expect(document.querySelector(".cm-editor")).not.toBeInTheDocument();
@@ -136,14 +223,14 @@ describe("Memo - edit mode", () => {
 
   test("saveMemo is not called in readOnly mode", async () => {
     vi.useFakeTimers();
-    render(Memo, { props: { saveMemo, content: "text", readOnly: true } });
+    renderMarkdownMemo({ saveMemo, content: "text", readOnly: true });
     vi.runAllTimers();
     expect(saveMemo).not.toHaveBeenCalled();
     vi.useRealTimers();
   });
 
   test("switching back to read mode without changes does not call saveMemo", async () => {
-    render(Memo, { props: { saveMemo, content: "hello" } });
+    renderMarkdownMemo({ saveMemo, content: "hello" });
     await fireEvent.click(document.querySelector(".edit-mode-btn"));
     await waitFor(() => expect(document.querySelector(".read-mode-btn")).toBeInTheDocument());
 
@@ -161,13 +248,11 @@ describe("Memo - edit mode", () => {
     const originalGetClientRects = Range.prototype.getClientRects;
     Range.prototype.getClientRects = () => [];
 
-    render(Memo, {
-      props: {
-        saveMemo,
-        content: "",
-        workspaceProjectDir: "C:\\project",
-        taskId: "task-1",
-      },
+    renderMarkdownMemo({
+      saveMemo,
+      content: "",
+      workspaceProjectDir: "C:\\project",
+      taskId: "task-1",
     });
 
     await fireEvent.click(document.querySelector(".edit-mode-btn"));
@@ -207,54 +292,13 @@ describe("Memo - edit mode", () => {
 
     Range.prototype.getClientRects = originalGetClientRects;
   });
-
-  test("pasting an image in db.json mode inserts a data URL image", async () => {
-    const originalGetClientRects = Range.prototype.getClientRects;
-    Range.prototype.getClientRects = () => [];
-
-    render(Memo, {
-      props: {
-        saveMemo,
-        content: "",
-        taskId: "task-1",
-      },
-    });
-
-    await fireEvent.click(document.querySelector(".edit-mode-btn"));
-    await waitFor(() => {
-      expect(document.querySelector(".cm-editor")).toBeInTheDocument();
-    });
-
-    const file = new File(["image-bytes"], "pasted-image.png", { type: "image/png" });
-    const pasteEvent = new Event("paste", { bubbles: true, cancelable: true });
-    Object.defineProperty(pasteEvent, "clipboardData", {
-      value: {
-        items: [
-          {
-            type: "image/png",
-            getAsFile: () => file,
-          },
-        ],
-      },
-    });
-
-    document.querySelector(".cm-content").dispatchEvent(pasteEvent);
-
-    await waitFor(() => {
-      expect(document.querySelector(".cm-content").textContent).toContain(
-        "![pasted-image](data:image/png;base64,"
-      );
-    });
-    expect(window.electronAPI.wsSaveMemoImage).not.toHaveBeenCalled();
-
-    Range.prototype.getClientRects = originalGetClientRects;
-  });
 });
 
-describe("Memo - link handling in preview", () => {
+describe("Markdown Memo - link handling in preview", () => {
   let saveMemo;
 
   beforeEach(() => {
+    quillInstances.length = 0;
     saveMemo = vi.fn();
     window.electronAPI = { openExternalLink: vi.fn(), wsResolveMemoAsset: vi.fn() };
   });
@@ -264,7 +308,7 @@ describe("Memo - link handling in preview", () => {
   });
 
   test("clicking a markdown link opens it externally and does not enter edit mode", async () => {
-    render(Memo, { props: { saveMemo, content: "[Visit](https://example.com)" } });
+    renderMarkdownMemo({ saveMemo, content: "[Visit](https://example.com)" });
     const link = document.querySelector(".preview a");
     expect(link).toBeInTheDocument();
 
@@ -279,13 +323,11 @@ describe("Memo - link handling in preview", () => {
 
   test("clicking a wiki link opens the target memo and does not enter edit mode", async () => {
     const openMemoLink = vi.fn();
-    render(Memo, {
-      props: {
-        saveMemo,
-        content: "[[Research Notes]]",
-        memoTitles: ["Research Notes"],
-        openMemoLink,
-      },
+    renderMarkdownMemo({
+      saveMemo,
+      content: "[[Research Notes]]",
+      memoTitles: ["Research Notes"],
+      openMemoLink,
     });
 
     const link = document.querySelector(".preview a[data-memo-link]");
@@ -300,7 +342,7 @@ describe("Memo - link handling in preview", () => {
   });
 
   test("clicking an external wiki link opens it externally", async () => {
-    render(Memo, { props: { saveMemo, content: "[[https://example.com|Example]]" } });
+    renderMarkdownMemo({ saveMemo, content: "[[https://example.com|Example]]" });
     const link = document.querySelector(".preview a");
     expect(link).toBeInTheDocument();
 
@@ -311,9 +353,10 @@ describe("Memo - link handling in preview", () => {
   });
 
   test("clicking non-link preview area stays in read mode", async () => {
-    render(Memo, { props: { saveMemo, content: "plain text" } });
+    renderMarkdownMemo({ saveMemo, content: "plain text" });
     await fireEvent.click(document.querySelector(".preview-mode"));
     await tick();
     expect(document.querySelector(".cm-editor")).not.toBeInTheDocument();
   });
 });
+
